Quefrency rev1: Fix default VIA layout options (#9618)
Apparently VIA allocates bits in the layout options field from the
lowest bit, but starting from the **last** option defined in the JSON
file.  So the default value 0x06 was actually trying to set the value
`3` (`0b11`) for the second-to-last option ("Right Shift"), which had
only 3 values defined, and the attempt to set an undefined option value
caused the VIA app to hang with a black window.

Fix the default layout options so that it works as intended (the
"Macropad" and "65% Column" options are set).